refactor: split the oversized downloader module into focused submodules
downloader.rs had grown to ~2.7k lines and mixed several unrelated
concerns, which made the module hard to navigate. Break it into a
`downloader/` directory of single-purpose submodules:

- signal: Ctrl+C handling (setup_signal_handler, handle_ctrl_c)
- retry: backoff/jitter, RetryTracker, parse_retry_after
- verify: size + MD5 checks (calculate_md5, check_existing_file,
  verify_downloaded_file)
- stream: response-body streaming, resume and download_file_content
- mtime: Last-Modified / _files.xml mtime handling

mod.rs keeps the batch orchestration (DownloadTask, download_files),
the .part setup/install helpers, and the public API (parse_last_modified,
sync_file_mtime re-exported). Shared helpers move to their natural home:
finish_progress_bar to utils, md5_hex/test_running to test_support.

Pure restructuring — no behaviour change: the public API is unchanged
(main.rs and archive_metadata.rs need no edits) and the test suite is
unchanged (103 passing), with unit tests co-located in their submodules
and integration tests kept beside the orchestration in mod.rs.
